refactor(corehr): auditoría arquitectónica, persistencia robusta y alineación OpenAPI
Resumen de Cambios
Refactorización profunda del módulo CoreHR para resolver deuda técnica crítica, estandarizar tipos de datos (Long), alinear la capa de adaptadores con la fuente de verdad (OpenAPI) y blindar la lógica en la capa de Dominio.

1. Arquitectura Hexagonal y Dominio

Estandarización de Identificadores: Se refactorizó TenantUserPort para operar estrictamente con Long. Se eliminaron casteos innecesarios (String.valueOf) en los consumidores.

Persistencia JDBC Precisa: Reescritura total de TenantUserJdbcAdapter para consultar correctamente desde tenant_user (JOIN con users/persons), mapeando atributos omitidos como phone y garantizando la resolución del state.

Dominio Rico en Asistencias (Attendance.java): Se reemplazó la recreación del agregado por comportamiento real. Ahora updateCheckIn y updateCheckOut disparan touch() internamente, asegurando la consistencia temporal.

Integridad Biométrica (UserBiometria.java): Extensión del modelo de dominio para incorporar y persistir templateCifrado y llaveCifrado.

2. Alineación Contractual (OpenAPI)

Filtros Funcionales: Se conectaron los parámetros de consulta ignorados (branch_id, area_id, departamento_id, tenant_user_id, rangos y estado) directamente al AttendanceRepositoryAdapter. El parámetro scope se mantiene como passthrough.

Contratos de Contacto: Corrección de TenantContactResult y TenantContactResponse para exponer el atributo phone (previamente en nulo) y asegurar la consistencia del ID como Long.

Privacidad de Biometría: El payload cifrado ahora se persiste correctamente mediante el EnrollService y su Mapper, asegurando que los datos sensibles jamás se expongan en las respuestas GET.

3. Integración Inter-Módulos (Modulith)

Módulo reports: Alineación end-to-end del tenantUserId a Long en puertos de salida, dominio (MonthlyClosureDetail) y adaptadores. Se corrigió un bug de truncamiento en CoreHrAttendanceSummaryJdbcAdapter aplicando rs.getLong("id").

Módulo incidents: Actualización en IncidentService para consumir el nuevo flujo, agregando tolerancia de parseo para mantener retrocompatibilidad con IDs legacy no numéricos.

Límites Arquitectónicos: Se validó que la comunicación cross-módulo se mantenga 100% aislada a través de Puertos de integración.

4. Pruebas y Quality Gate (SonarQube)

Cobertura Ampliada: Se implementaron pruebas exhaustivas para TenantUserJdbcAdapter, TenantContactService, flujos de cifrado biométrico, y casos de borde en correcciones de asistencia.

Estado de la Suite: 1056 tests ejecutados | 0 fallos | 0 errores.

5. Compilación y Deuda Técnica

Build exitoso y limpio al ejecutar mvn clean verify.

Nota Técnica: Se documenta para el backlog un Deprecation Warning en SecurityConfig.java relacionado con AntPathRequestMatcher (no bloqueante para esta entrega).

+++ back/src/main/java/trazzo/back/corehr/application/port/in/CoreHrAttendanceSummaryPort.java
@@ -7,7 +7,7 @@ public interface CoreHrAttendanceSummaryPort {
     List<EmployeeMonthlySummary> getMonthlySummaries(int month, int year);
 
     record EmployeeMonthlySummary(
-            Integer tenantUserId,
+            Long tenantUserId,
             String tenantUserFullName,
             String tenantUserDocument,
             String departmentName,
