+++ back/src/main/java/trazzo/back/reports/infrastructure/adapters/in/web/dto/MonthlyClosureDetailResponse.java
@@ -4,7 +4,7 @@ import java.time.LocalDateTime;
 import java.util.UUID;
 
 public record MonthlyClosureDetailResponse(
-        UUID id, UUID monthClosureId, Integer tenantUserId,
+        UUID id, UUID monthClosureId, Long tenantUserId,
         String tenantUserFullName, String tenantUserDocument,
         String departmentName, String roleName,
         Double totalWorkedHours, Integer totalTardinessMinutes,
